+++ appseed/aura/aura/os/macos/macos_process_pid.cpp
@@ -93,7 +93,7 @@ int_array get_pids()
    
    pida.set_size(numberOfProcesses);
    
-   proc_listpids(PROC_ALL_PIDS, 0, pida.get_data(), pida.get_size());
+   proc_listpids(PROC_ALL_PIDS, 0, pida.get_data(), (int) (pida.get_size()));
    
    for(auto pid : pida)
    {
@@ -509,7 +509,7 @@ int get_process_pid(const char * csProcessName)
 int get_current_process_affinity_order()
 {
    
-   int numCPU = sysconf(_SC_NPROCESSORS_ONLN);
+   int numCPU = (int) (sysconf(_SC_NPROCESSORS_ONLN));
    
    return numCPU;
 }
